+++ AAEmu.Game/Core/Packets/C2G/CSChangeHousePermissionPacket.cs
@@ -13,7 +13,7 @@ public class CSChangeHousePermissionPacket : GamePacket
 
     public override void Read(PacketStream stream)
     {
-        var tl = stream.ReadUInt16();
+        var tl = (ushort)stream.ReadUInt32(); // the handle is 32 bits on the wire; ours is 16
         var permission = stream.ReadByte();
 
         Logger.Debug("ChangeHousePermission, Tl: {0}, Permission: {1}", tl, permission);
